CONTRIB-8600: Move completion to custom_completion class
